% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{athlytics_colors_ef}
\alias{athlytics_colors_ef}
\title{Efficiency Factor Colors}
\usage{
athlytics_colors_ef()
}
\value{
A named list with four color codes by sport:
\item{run}{Navy blue for running}
\item{ride}{Coral for cycling}
\item{swim}{Cyan for swimming}
\item{other}{Slate for other activities}
}
\description{
Colors for efficiency factor trends by activity type.
}
\examples{
# Get EF colors by sport
colors <- athlytics_colors_ef()
colors$run # Navy for running

}
