% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{athlytics_colors_training_load}
\alias{athlytics_colors_training_load}
\title{Training Load Colors}
\usage{
athlytics_colors_training_load()
}
\description{
Colors for acute and chronic training load visualization.
}
