% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{athlytics_palette_science}
\alias{athlytics_palette_science}
\title{Science Magazine Palette}
\usage{
athlytics_palette_science()
}
\description{
Classic palette inspired by Science magazine's figure guidelines.
Conservative and widely accepted in scientific community.
}
