% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{athlytics_sample_decoupling}
\alias{athlytics_sample_decoupling}
\title{Sample Aerobic Decoupling Data for athlytics}
\format{
A tibble with X rows and 2 variables:
\describe{
  \item{date}{Date of the activity, as a Date object.}
  \item{decoupling}{Calculated decoupling percentage, as a numeric value.}
}
}
\source{
Simulated data generated for package examples.
}
\usage{
athlytics_sample_decoupling
}
\description{
A dataset containing pre-calculated aerobic decoupling percentages,
derived from simulated Strava data. Used in examples and tests.
}
\keyword{datasets}
