% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{athlytics_sample_exposure}
\alias{athlytics_sample_exposure}
\title{Sample Training Load Exposure Data for Athlytics}
\format{
A tibble with X rows and 5 variables:
\describe{
  \item{date}{Date of the metrics, as a Date object.}
  \item{daily_load}{Calculated daily training load, as a numeric value.}
  \item{ctl}{Chronic Training Load, as a numeric value.}
  \item{atl}{Acute Training Load, as a numeric value.}
  \item{acwr}{Acute:Chronic Workload Ratio, as a numeric value.}
}
}
\source{
Simulated data generated for package examples.
}
\usage{
athlytics_sample_exposure
}
\description{
This dataset contains daily training load, ATL, CTL, and ACWR, derived from
simulated Strava data. Used in examples and tests, particularly for `plot_exposure`.
}
\keyword{datasets}
