% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_decoupling.R
\name{calculate_pace_hr_decoupling}
\alias{calculate_pace_hr_decoupling}
\title{Calculate Pace/HR Decoupling for Running}
\usage{
calculate_pace_hr_decoupling(stream_data)
}
\description{
Calculate Pace/HR Decoupling for Running
}
\keyword{internal}
