% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_decoupling.R
\name{calculate_power_hr_decoupling}
\alias{calculate_power_hr_decoupling}
\title{Calculate Power/HR Decoupling for Cycling}
\usage{
calculate_power_hr_decoupling(stream_data)
}
\description{
Calculate Power/HR Decoupling for Cycling
}
\keyword{internal}
