% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{athlytics_palette_science}
\alias{athlytics_palette_science}
\title{Science Magazine Palette}
\usage{
athlytics_palette_science()
}
\value{
A character vector of 8 hex color codes
}
\description{
Classic palette inspired by Science magazine's figure guidelines.
Conservative and widely accepted in scientific community.
}
\examples{
# Get Science palette colors
colors <- athlytics_palette_science()
colors[1]  # Dark blue

}
