% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pbs.R
\name{find_best_effort}
\alias{find_best_effort}
\title{Find Best Effort for Target Distance}
\usage{
find_best_effort(stream_data, target_distance)
}
\description{
Find Best Effort for Target Distance
}
\keyword{internal}
